+++ packages/react/examples/codesandbox/components/ContentItemHorizontal/src/index.js
@@ -1,5 +1,5 @@
 /**
- * Copyright IBM Corp. 2016, 2020
+ * Copyright IBM Corp. 2016, 2022
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
@@ -7,8 +7,7 @@
 
 import './styles.scss';
 
-import { ContentItemHorizontal } from '@carbon/ibmdotcom-react';
-import React from 'react';
+import { ContentItemHorizontal } from '@carbon/ibmdotcom-react/es/components/ContentItemHorizontal';
 import ReactDom from 'react-dom';
 
 const App = () => (