+++ packages/web-components/src/components/content-block-cards/content-block-cards.ts
@@ -43,4 +43,5 @@ class DDSContentBlockCards extends StableSelectorMixin(DDSContentBlock) {
   }
 }
 
+/* @__GENERATE_REACT_CUSTOM_ELEMENT_TYPE__ */
 export default DDSContentBlockCards;