+++ packages/services/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.37.0](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/compare/@carbon/ibmdotcom-services@1.37.0-rc.1...@carbon/ibmdotcom-services@1.37.0) (2022-08-17)
+
+**Note:** Version bump only for package @carbon/ibmdotcom-services
+
+
+
+
+
+# [1.37.0-rc.1](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/compare/@carbon/ibmdotcom-services@1.37.0-rc.0...@carbon/ibmdotcom-services@1.37.0-rc.1) (2022-08-05)
+
+**Note:** Version bump only for package @carbon/ibmdotcom-services
+
+
+
+
+
+# [1.37.0-rc.0](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/compare/@carbon/ibmdotcom-services@1.36.0...@carbon/ibmdotcom-services@1.37.0-rc.0) (2022-07-29)
+
+
+
+## 1.35.1 (2022-07-08)
+
+**Note:** Version bump only for package @carbon/ibmdotcom-services
+
+
+
+
+
 # [1.36.0](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/compare/@carbon/ibmdotcom-services@1.36.0-rc.0...@carbon/ibmdotcom-services@1.36.0) (2022-07-18)
 
 **Note:** Version bump only for package @carbon/ibmdotcom-services